requirements.txt
asgiref==3.5.2
autopep8==2.0.0
black==22.10.0
click==8.1.3
Django==4.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.5
sqlparse==0.4.3
tomli==2.0.1
